dev env: fix deprecated KillMode=none
These units were originally created by "podman generate systemd"
which inserted a KillMode=none directive. That value of KillMode
was deprecated by systemd and will be removed in the future.

This commit updates the units to be aligned with podman's fix for
the issue in https://github.com/containers/podman/issues/8615,
which was to drop KillMode and add TimeoutStopSec.