Improve Publish model's resolve_links [RHELDST-11634]
Previously this method loaded every item on the publish, storing them in
memory for its duration. The number of these items can realistically
reach into many hundreds of thousands, approaching a million.

This commit refactors the method to use a querying approach to store
only the items that it needs to operate on -- those with non-empty
link_to fields -- and load only the URI and object key of the
corresponding items. These changes should significantly reduce the
impact this method has on system memory.