ci: fix compatibility with pylint>=2.14
pylint 2.14 introduced more strict argument parsing. Enabling or
disabling a nonexistent check now generates a warning, which causes
"tox -e static" to fail. Clean up the list of checks to contain only
those which really exist as of pylint 2.14.

For this repo we also invert the behavior of "enable all pylint checks,
then disable undesirable checks" to instead "disable all pylint checks,
then enable desirable checks". This is more consistent with other repos
and generally results in a more stable CI.

The set of enabled checks are not precisely the same as before, so some
minor tweaks in the code were needed to get it passing.