Fix logging config not respected by worker [RHELDST-12333]
The Settings class takes settings from two sources: environment
variables and exodus-gw.ini. All logging config lives in the INI
file.

Problem: the proper loading of settings and initialization of loggers
was hooked into the WSGI app startup and not used for the worker
component, which would simply construct a 'Settings()'. This meant any
non-env-var config would be silently missed by the worker, including
configuration of the loggers.

Hence, adjusting the log level would have no effect for the worker process.

Fix it by ensuring that the worker loads settings properly and
initializes loggers on startup. The logger initialization shares code
with the web component. A minor change to .ini file loading was also
added to improve testability, as it was not previously possible to make
a test point at a specific config file.