dev env: tweak suggested awscli commands
Because no profile was specified, the commands as documented here might
not actually work - it would depend on the state of the caller's AWS
config files.

We know that a profile named 'test' should exist because
scripts/localstack-init creates one if needed, so let's adjust the
commands to refer to it. It should make the commands a bit more reliable
across environments.