dev env: tweak exodus-gw.ini cdn_url for fakefront
The test env in this exodus-gw.ini is primarily used for the development
environment. Rather than pointing it at a bogus cloudfront URL which will
never work, let's point it at the fakefront instance deployed in that
development environment so the CDN redirect endpoints have a chance at
working out of the box.