logging: prefix log messages with current actor
Add a new middleware which will arrange for all log messages produced
within an actor to be prefixed with at least the actor's name, plus
task/publish ID being processed where possible.

This ensures that every log message can be traced back to a particular
publish/task which makes it much easier to quickly find relevant logs,
without requiring the logging statements to explicitly log the task
being handled.

Here's an example of the impact of this change:

-Items successfully written
+[commit c3234df2-cdaf-4df1-aa1e-b360fd68f8a1] Items successfully written

This change was done in preparation for RHELDST-12449 which has some
requirements around logging.