exodus-gw.ini: use more appropriate log levels
This config file is used both when running tests and from the exodus-gw
dev env.

The log levels requested here were presumably arbitrary for testing,
but they don't make sense for the dev env or in general: they set the
exodus-gw logger to *less* verbose than all other loggers, which is the
opposite of what we'd want in practice.

Flip it around so the root and exodus-gw loggers use more appropriate
levels (also matching the config we actually use for deployment), and
fix up the one affected test.