Parallelize batch writes [RHELDST-13451]
This commit ensures that DynamoDB clients are not unnecessarily
created/destroyed repeatedly during the commit. It also ensures
that DynamoDB batch writes happen concurrently. The number of threads
is configurable in the server settings via "write_max_workers".

This commit also includes refactoring to make the concurrent batch
writes a bit cleaner.