Improve use of batch write threads [RHELDST-13926]
Previously, when writing items to DynamoDB, a single partition would be
loaded and sumbitted to the thread pool executor, blocking until all
items in the partition were completed.

This commit allows the number of partitions to be configured, meaning
blocking only occurs once all item batches within n number of partitions
are submitted to the executor.

Note that items are still only loaded per partition, so there should be
no drastic memory usage increase. There is, however, correlation between
number of partitions started and memory used due to the futures stored for
each batch write but this should not normally be significant.