dev-env: mitigate race condition during service startup
When starting up all units it seems that it was possible for the
following to happen:

- exodus-gw-db service is marked active, but hasn't yet started to
  accept connections

- exodus-gw-uvicorn starts up, fails to initialize because it can't
  connect to DB

- exodus-gw-db starts to accept connections; but because
  exodus-gw-uvicorn uses hot reload, it will not automatically exit
  and be restarted by systemd, so the service remains broken
  indefinitely (could be fixed by manually restarting that service)

In this commit we add a post-start command to exodus-gw-db which waits
for the expected TCP port to become open. It should make the above
less likely to happen as the startup of exodus-gw-uvicorn will be
delayed until the DB is apparently ready for use.