Avoid sqlalchemy 2.x for now [RHELDST-15252]
We already know that we aren't yet compatible with sqlalchemy 2.x
and filed an issue to work on that.

Given that sqlalchemy 2.0.0 was recently released, our weekly dependency
updates now try to make use of it and will fail CI. Request an older
compatible version until we fix the compatibility issues.