Fix mypy error with latest versions of dependencies
With latest versions of all packages, mypy has started to complain:

      exodus_gw/routers/publish.py:304: error: Missing positional argument
          "settings" in call to "send" of "Actor"  [call-arg]

There is no actual type error here - we have dramatiq middleware to
insert an instance of 'Settings' to any actor with a settings argument.
However, mypy is not aware of that.

It can be easily fixed by declaring a default, so just do that. In
practice we never expect the default to be used and this is only to help
mypy.