Fix some logging issues with BatchWriter [RHELDST-17783]
Firstly, assign a meaningful thread name so it's easier to tell when log
messages are coming from the BatchWriter.

Secondly, fix issues with the actor log message prefixing.

Log messages coming from within an actor are intended to be prefixed
with the actor's name and task ID, e.g. "[commit cf2f...a3e]".

However, this wasn't working from within the BatchWriter helper class.
The reason for this was that the logging middleware relied on a
thread-local to maintain the current log prefix, but BatchWriter
created new threads internally which would then see different values
on the thread-local.

It can be solved by switching from a thread-local to the more generic
contextvars.Context. It is possible to explicitly propagate a context
from one thread to another to share the log prefix (or any other
contextvars), fixing the problem.

One annoyance with this approach: if any new code would be added which
also spawns new threads, that code will have to remember to do the same
trick with copy_context. I can't come up with a way to make it work
automatically for all new code.