Fix thread-safety issues in DynamoDB helper class
Instances of this class are used from within multiple threads in
BatchWriter, but the code was not written to be thread-safe (racy
lazy initialization of some attributes). Fix it:

1) no lazy initialization for 'client'. If you're creating an instance of
   this class, you're definitely going to use the client, so there is no
   benefit to the added complexity of delaying initialization.

2) safe lazy initialization for 'definitions'. Not every usage of the
   class might read definitions, so this can be lazy, however we should
   use a lock to ensure it only happens once even when multiple threads
   are using the object.

The motivation for this change: while reviewing localstack logs during
commit operations, I found that there were many dynamodb Query
operations at the beginning of a commit. This doesn't seem right, as the
only query during commit is to load 'definitions', but that was intended
to happen only once. (2) ought to fix that. It could potentially give
a minor performance improvement on slower environments such as
localstack.