Log "login" events [RHELDST-19460]
Our security standards require us to log "login" events.

Although exodus-gw does not technically provide a "login" endpoint,
we have defined a login event as, "when platform-sidecar accepts a
presented cert as valid during TLS handshake."

Unfortunately, the platform-sidecar logs do not seem to indicate
whether a certificate has been provided with the request in question.

exodus-gw already produces an "auth" log containing the requested
path, username, and roles when an endpoint that requires authorization
has been requested. Though, exodus-gw also provides read-only endpoints
which do not require authorization. It technically possible for a user
to provide a cert (and for platform-sidecar to consider it valid) to
endpoints which do not require authorization. A log was previously not
produced in these instances.

Now, the user (or "anonymous user" if unauthenticated), roles (if any),
and the requested path are now logged with each request to comply with
our security standards. The path can be cross-referenced with the
platform-sidecar logs, if/when needed.