Add logging handler to write to healthcheck file [RHELDST-19614]
To track the healthiness of workers in OCP, we'll probe for a specific
file, checking it's latest update time and restarting should a certain
time elapse since that update.

This commit implements the necessary mechanism for writing that file.
By using a custom logging handler, we implement an emit method to log
updates the healthcheck file. The location of the healthcheck file can
be configured via a new setting, "worker_health_filepath".