Ensure openapi files are copied to docs [RHELDST-19887]
Previously we relied on Sphinx to overwrite build files with those
listed in html_extra_files, specifically api.html. However, for some
reason, that is no longer happening. This commit adds a copy command to
the gen-openapi script to ensure that the appropriate files are present
in the final build.