Update S3ClientWrapper to prevent retries [RHELDST-20355]
It was discovered that boto client configuration supports two retry
attempt settings. The one used previously is not inclusive of the initial
attempt so `"max_attempts": 1` means one retry will be attempted. This
commit changes this setting to "total_max_attempts" because it is 1)
preferred by the library and 2) inclusive of the initial attempt.

This directly addresses the intermittent UnseekableStreamErrors seen
recently. These errors are only raised when boto attempts to retry
S3 uploads, which we were unknowlingly commanding it to do.

Additionally in this commit, a key-value pair is added to the client's
request to prevent redirects if/when botocore's newest redirector class
is used.