refactor: move commit tasks into their own table
Previously the Task model had a publish_id column which was nullable,
because some tasks relate to specific publishes (e.g. commit) and others
do not (e.g. deploy-config).

For RHELDST-20490 I'd like to add some more columns which are specific
to commit, but it's not really a good practice to dump all type-specific
columns into a single table and allow them to be nullable.
Inheritance can be used to instead have Task as the base class and
CommitTask as a subclass adding additional fields, so let's do that.

The added migration includes usage of the recently added `tested_by`
helper to ensure that migration of existing tasks doesn't crash.
This also uncovered one (not very important) bug in an existing
migration's "downgrade".