Fix mismatch between sqlalchemy model and migrations
While working on one issue I noticed that "alembic autogen" was not
clean, i.e. it wanted to generate some migrations even though I'd not
changed the model, indicating some drift between the two.

There were two issues, both fixed, and in both cases treating the
migrations as correct and the model as the place to be fixed.

1. In cfa6812 (upgrade to sqlalchemy 2.0), the "nullable" state of
   various columns was unintentionally changed.

   Take tasks.deadline as an example. In our migrations, the column was
   always created as nullable=True. In the model, it was declared as a
   Mapped[datetime] with no specific 'nullable'; but per [A], that means
   it's derived from the type hint. Optional means nullable and
   non-Optional means not nullable, which makes sense. Hence that commit
   actually changed the column from nullable to not-nullable in the
   model, but didn't add any migration to do the same.

   While fixing this I also removed all 'nullable' from all columns so
   that we rely *solely* on the type hints; I think this is clearer than
   having two different ways of declaring whether something is nullable.

2. In 102733ca a migration was added to create a uniqueness constraint,
   but there was no attempt to add it to the model at all. Add the
   missing constraint now.

As well as the fixes, a new test was added which runs "alembic check"
to prevent similar mistakes in the future.

The addition of that test required fixing one other unrelated problem:
the JSONB hack in sqlite_compat was, for some reason, causing JSONB
columns to end up in the schema as NUMERIC. That doesn't actually break
sqlite since, as explained in the comment above, everything is text
in sqlite anyway; but, starting from alembic 1.12.0 which released about
a month ago, 'alembic check' detects this difference and so would fail
the test.

Everything works fine if the hack is changed to instead declare the
column as "TEXT"; the only reason it wasn't like that originally was
because it seemed clearer if the sqlite schema would display the column
as JSONB to indicate that it's not expected to contain arbitrary text.

[A] https://docs.sqlalchemy.org/en/20/orm/mapping_api.html#sqlalchemy.orm.mapped_column.params.nullable