Set from_date from item update time [RHELDST-20491]
All items written to DynamoDB are sorted by from_date.

Up to now, we have always taken the from_date from the time at which the
commit task was enqueued. This means if two publishes would try to
update the same path at the same time, whoever commits last "wins".

Those semantics are in fact not the most useful, which was exposed once
realistic Pulp usage occurred. The problem is that it allows a sequence
like the following:

1. Repo publishes within exodus-gw publish A.

2. Repo publishes again, within exodus-gw publish B.

3. Publish B is committed.
  => new state: repodata from (2) is available on CDN
     (which also contains packages added at step (1))

4. Publish A is committed.
  => new state: repodata from (1) is available on CDN.
     BUG! repodata has effectively gone backwards, "unpublishing"
     whatever was added in step (2).

It makes more sense to take the from_date from the time at which each
individual item was updated. This makes the above scenario work as
desired:

1. Repo publishes within exodus-gw publish A.

2. Repo publishes again, within exodus-gw publish B.

3. Publish B is committed.
  => new state: repodata from (2) is available on CDN.
     (which also contains packages added at step (1))

4. Publish A is committed.
  => inserts DynamoDB items for the repodata from (1), but since these
     have an earlier from_date than what's already in the DB, there is
     effectively no impact on the repodata.

This change ensures that, if multiple publishes are updating the same
path, everything keeps moving strictly forward even if commits are done
in the reverse order from updates.

Note that the commit actor still retains a 'from_date' because that's
also used to query exodus-config while running, and it's still
reasonable to use the commit time for that.