autoindex: don't re-upload existing objects [RHELDST-20661]
This code was missing the usual logic to do a HEAD request to check
whether an object was already in S3 before uploading it. Add the
missing logic.

This is important since object versioning is enabled, so each upload
would create a new object version and incur an ongoing cost. In
practice, indexes across many repos generate an identical object (e.g.
the top level of every yum repo looks the same - just "Packages" and
"repodata"), so the number of unwanted object versions created due to
this is quite high. The aforementioned top level yum repo index object
already has ~9000 or so versions in production.