Fix tasks not promptly set to FAILED in some cases [RHELDST-20663]
The responsibility of updating a Task object to completed or failed
generally lies within the dramatiq actor (e.g. the commit actor itself
updates its corresponding Task).

However, there are some cases where that might not happen. One example
is when a commit task fails at the autoindex step, as this happens
before the main DynamoDB write step wrapped in try/finally. Another
example would be openshift-level problems such as worker pods being
killed with OOM in the middle of a task.

In such cases, a task may be left IN_PROGRESS even after a message
failed to process. Our scheduled Janitor would eventually detect and fix
that, but this is only intended to run every 12 hours or so, which is
unreasonably long for e.g. Pub to learn that a task has failed.

To address this, at the same place where we delete a DramatiqMessage
after all retries were exhausted, we should also mark any corresponding
incomplete Task as failed. This will allow the failure to be observed
promptly rather than having to wait for the next Janitor run.