Pass S3 error codes back to client [RHELDST-20675]
Any error responses from S3 would be previously caught by our generic
exception handler and turned into a 500 internal server error.

We shouldn't do that, instead we should be giving the real error codes
from upstream S3 back to the exodus-gw client so that:

- if the error is the client's fault, they can see the reason for that
  (e.g. BadDigest)

- if the error is of a type expected to happen during the course of
  normal S3 usage, make sure the client can see that error properly
  so that they can retry as normal (e.g. SlowDown)

- errors from S3 don't drag down the SLI of our service

This commit adds a new exception handler which passes back any upload
errors using the proper status and error response.

Note that head_object already had special handling for this,
specifically for 404 errors, since we've always used that to determine
whether an object exists. There is no need for that to have special
handling, as the desired behavior (pass back any errors from S3) is the
same for all of the upload endpoints. Hence that special handling has
been removed, along with an associated test.