docs: tweak language regarding items on "Get Publish"
This API doesn't load/return the "items" list on a publish. The way this
was explained before was not appropriate for a user-oriented doc.

The reference to noload() in particular is a very inappropriate detail
since a user of the exodus-gw API does not need to know or care that
SQLAlchemy is used, or even that an SQL-based DB is used.

(We probably should have made this API just not return the 'items'
element at all, instead of an empty list, but it's a bit late for that
now.)