Fix phase1 concurrency issue leading to empty object_key [RHELDST-21776]
For symlink items (i.e. those with "link_to" set), links are resolved at
the time a commit is requested.

Previously it was possible for this to occur:

1. Client A adds items to a publish.

2. Client A enqueues phase1 commit (resolving links).

3. Client B adds more items onto publish, including some more links.

4. Worker processes the phase1 commit enqueued in (2).

   BUG: the commit processes the items added at (3) even though links
   have not been resolved for those, meaning they have an empty
   "object_key".

Writing items with an empty object_key would ultimately result in
exodus-lambda redirecting to the root of the S3 bucket, giving 403
errors.

This commit fixes the issue by making phase 1 commit skip the processing
of items with a missing object_key, which are assumed to be
not-yet-resolved link items. Phase 2 commit still processes all items as
before.

Additionally, a check was added to ensure that items have an object_key
prior to write. Although there is no known way to fail that check now
that this bug is fixed, it nevertheless seems worth an explicit check in
case of similar issues in the future.