dev-env: fix paths for Fedora 39
The meaning of "%S" was changed between systemd versions in F38 vs F39;
see [1] vs [2]. For a user manager it was previously equal to
$XDG_CONFIG_HOME (~/.config), then it was changed to $XDG_STATE_HOME
(~/.local/state).

There is nothing inherently wrong with these units using the state dir,
but some of our docs and scripts assume that the path is ~/.config. It's
also difficult to support F38 and F39 at the same time, or support an
upgrade from F38 to F39, if both paths are in use.

So, switch to "%E" instead, which consistently uses ~/.config across
systemd versions.

[1] https://www.freedesktop.org/software/systemd/man/247/systemd.unit.html
[2] https://www.freedesktop.org/software/systemd/man/latest/systemd.unit.html