Fix link resolution not isolated to current publish [RHELDST-21893]
Items using link_to should only be resolved to a proper object within
the current publish. If resolution is not limited in this way, it's even
possible for links set up on "pre" to leak over into "live" or
vice-versa.

Although link resolution was always intended to work this way since
initially scoped, the fact that it was buggy ever since implemented
means that unfortunately other systems might have come to rely on it,
and Pub/Pulp in the "hook disabled" case is known to rely on it.
For that reason, a temporary escape hatch is added to allow continuing
with the original buggy behavior if needed.

There is no specific test in this repo targeting this bug, but there is
a system test in pipeline-test repo covering it.