Fix crash during commit if attempting to delete an entrypoint
The special object key of "absent" is used on an item to effectively
delete it. The autoindex code didn't filter these out, so when it saw
e.g. a repodata/repomd.xml having key "absent", it would really try to
fetch that object and generate an index, causing commit to fail.