docs: hide useless version
We'd set a version of 1.0.0 here, and the idea is that this should be
updated as API features are added and removed, but in practice we never
bother doing that.

The version appears in the docs and it's misleading to have a version of
1.0.0 always displayed there even as the API changes. Since we aren't
attempting to maintain this, just set a blank version instead. This
hides the version from the redoc-based API docs.