Fix inability to delete index pages [RHELDST-22047]
Since exodus-gw internally generates indexes automatically during
publish, the filename used for those is reserved to prevent conflicts
from clients.

However, by not allowing the usage of that filename when setting an
object to "absent", we make it impossible to *delete* any such generated
indexes. We need some way to delete them, so lift the restriction for
this particular object key.

Given that this requires the client to know the filename of the
generated indexes, which is supposed to be a hidden implementation
detail, the fix is not ideal. But some fix is needed quickly because we
otherwise cannot fully roll back any publishes performed via exodus-gw,
which we might need to do in case of issues. I'd like to return to this
later and find some better way of supporting indexes in the API.