docs: redirect from root of service to /redoc
Since there is no API at "/", previously if opening up the service in a
browser, one would see a 404 Not Found error in JSON format. It gives
the impression that there might be something wrong with the service.

Let's make it so that accessing the root of the service will redirect to
the docs to fix this and help with discoverability of the API.

This logic is limited to browsers (judged by clients requesting an HTML
response) so that other clients wouldn't come to depend on it.