Fix links repeatedly resolved for large publishes [RHELDST-22081]
After the introduction of phase1 commits, for one publish involving
(say) 600 Pulp repos, it's expected that a phase1 commit happens 600
times. Resolution of "link_to" happens as a part of this.

Problem: the query finding items for link resolution would find all
items with a non-empty "link_to", but we never *set* the field to empty
after resolving links. That means that every request to the commit
endpoint for that publish is resolving not only any recently added
links, but also all previously added links even if they were already
resolved before.

Worse, all the items selected for resolution need to be locked
FOR UPDATE, while any workers processing a commit *also* need to lock
the same items FOR UPDATE.

As a result, for publishes involving many repos, the commit endpoint
could slow down unreasonably over time.

This commit fixes it by wiping out the link_to field once we have
resolved links so that the same item does not get resolved more than
once.