Improve performance via partial link resolution [RHELDST-22147]
Links were previously resolved only at commit time. In most cases,
we have all the needed info to resolve links as items are being added
to a publish:

- if items being added are links to items which already exist in the DB,
  they can be immediately resolved before saving.

- if items being added are targets of links which already exist in the
  DB, those existing items in the DB can be updated as part of the
  current request.

This change adds the needed logic to resolve links during update for the
above two cases. The motivation is that, per observations during a very
large publish in RHELDST-22147, it was seen that some requests to
"/commit" took as long as 60 seconds. This probably indicates that we're
trying to do too much work in the context of a single HTTP request and
it's better to find a way to split up the work across multiple requests.

Note: because this commit added a type hint to the signature of
resolve_links, mypy started to type-check that method which was
previously not the case. For this reason the commit had to also fix up
some existing type issues in the method.