performance: use faster method of item insert
Adding items onto publishes seemed a bit slower than expected.
After some profiling and experimenting I found that changing the method
of bulk insert here significantly improves performance.

The sqlalchemy docs hint at this. In the docs for Insert.values[1]
it's mentioned that:

> To emit an INSERT statement against multiple rows, the normal method
> is to pass a multiple values list to the Connection.execute() method,
> which is supported by all database backends and is generally more
> efficient for a very large number of parameters.

The included test takes 15 seconds on my system to run prior to this
change and 6 seconds afterward.

[1] https://docs.sqlalchemy.org/en/20/core/dml.html#sqlalchemy.sql.expression.Insert.values