cleanup: drop unused LINK_ISOLATION hack
This conditional was added as an escape hatch just in case the bug fix
for RHELDST-21893 broke some exodus-gw users and would have to be
temporarily disabled. In practice we never needed to use this and it can
be cleaned up. Link isolation should always be enabled, as originally
intended.