Add missing URL encoding during CDN redirect [RHELDST-22493]
The URLs used here for redirect ought to be percent-encoded.

Although neither CloudFront nor exodus-lambda care about this, if not
encoded, some clients will break. python-requests is an example of a
client which breaks (while curl is OK).

What happens with python-requests is like this:

1. python-requests caller requests the CDN redirect API with a path
   having a reserved character, e.g. /some^path
2. python-requests automatically encodes to /some%5Epath
3. starlette/gunicorn/whatever automatically decodes to /some^path, so
   that's what the exodus-gw code sees
4. exodus-gw responds with a redirect, with Location: <cdn_endpoint>/some^path
5. python-requests automatically encodes *that* to <cdn_endpoint>/some%5Epath
   when issuing the follow-up request
6. since the final URL for the request used "%5E", but the URL used
   internally by exodus-gw when generating CloudFront signatures used
   "^", the request will fail CloudFront signature verification.

Quoting/escaping before signing should fix the issue, as the path
returned at step (4) would already use %5E, would be consistent with the
signature, and wouldn't be rewritten by python-requests or other
clients.