Add cache flush endpoint [RHELDST-23276]
This commit introduces a new endpoint, "/{env}/cdn-flush". It can be
used to explicitly perform cache flushing of the CDN edge when needed,
using Akamai Fast Purge API.

The intent is for this API to be used only for troubleshooting and for
niche scenarios. Later commits should introduce *implicit* cache
flushing into the usual publish workflow, so it would generally be
unnecessary to use this.